requirements.txt
# LLM API
groq>=0.4.0

# LangChain (optional)
langchain>=0.1.0
langchain-groq>=0.0.1

# Data handling
pandas>=2.0.0
datasets>=2.14.0

# Analysis and visualization
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
notebook>=7.0.0

# Code parsing and analysis
# ast is built-in, no need to install

# Testing and utilities
pytest>=7.4.0
pytest-timeout>=2.1.0

# Environment management
python-dotenv>=1.0.0

# Optional: for better progress tracking
tqdm>=4.66.0

# Optional: for prettier output
rich>=13.0.0
